moved controls and file_list to gui.controls
